Remove onSubmitHandler prop from Basic button component
I have added eventDefault back into the onClickHandler, so the onSubmitHandler should never be called anyway. Anything that can be done with the onSubmitHandler should be possible just by using the onClickHandler. I have also made the click event in the onClickHandler type a requirement to help with type errors in other components using the Basic button component - it was always being provided to the callback function anyway so there was no reason for it to be optional.