Added onBlur and autoFocus to chat Header input element
Now, when the input element is mounted, it automatically gains focus, and when it loses focus, it is removed. I know autoFocus isn't a very accessible practise, but I think in this situation it benefits the user experienc; clicking the 'Edit' button giving focus to the input element immediately makes sense here.