Change routes
I've changed how the project is structured using routes - before, when switching from / to /chats, the sidebar was re-rendering even though it was identical in both pages' components. To resolve this, I have placed the sidebar in the Root page component, and have implemented the Outlet component to render the rest of the Home and Chat pages' content within the Root component. Now, when switching between / and /chats, the sidebar does not re-render, as the Root component itself does not re-render as both paths are child paths of the path that renders the Root component.