Update route structure
The routes are now exported from the same component on which they are to be 'applied'. For example, if the Home page component needs to render 2 different elements based on the current 'nesting' of the URL, those routes are to be exported from the Home page component's file as a non-default export. Any component that then renders the component can then also import its routes for use.