Set up mock data in 'Main' profile component
I have also found a workaround for ensuring the 'selected' prop passed to the Horizontal navigation component from the 'Main' profile component is correct based on the current URL path. Before, whenever the page was refreshed, it would just default to whatever value was specified for the prop even if it didn't accurately reflect the current URL path.