Implement pub/sub pattern in components
The Root page component now subscribes to a topic that is published by the post List component when the 'Create New Post' button is clicked - this opens a modal. I am using the modal in the Root page component so it can fill the entire viewport; this wouldn't be possible using the modal in the post List component.