Fix problem in Basic form component
Due to the keys provided to each section element (using uuidv4), each time the component rerendered, the inputs were being reset to their initial values. For example, when changing the value of an input that had a validator, if the validation failed, it would possibly trigger a rerender in the form component depending on the state of the submission button, which would then rerender each component. Now, each section's key is either the section name or its index in the array (as I do not expect my forms to change), which causes them to persist through rerenders, leaving the inputs' updated values intact.