Update Basic modal component's positioning style rules
Before, when the containing element was too small, the modal within would overflow the container at the top which, when the containing element is the viewport, made the top of the modal inaccessible - this made the 'close' button impossible to click manually. This has now been fixed; the modal should not overflow the top of the containing element (by default).