Add extendedTypes file to shared package utils folder
I have also written a new function in the file to check if a given value is an instance of one of the options of the TypedArray union type. I will be removing the extendedTypes file from the frontend package utils folder in an upcoming commit to remove code duplication.