Add 'maximumAmount' prop to File component
The component will now disable the upload button when the current number of files is equal to the maximum amount specified in the prop. Also, trying to upload multiple files at once which would normally exceed the maximum amount specified will not work; the excess files will not be added.