Update layout for 'Main' profile component
Before, the posts/replies etc. underneath the profile summary were within a scrollable list, but on smaller screens, this part of the page would overflow off the screen with no way of viewing it. Now, the entire 'Main' profile component is within a scrollable list, so any part of it can be scrolled to, making it more accessible for users with devices with smaller screen sizes.