Overhaul application layout
Instead of the middle column on most pages being independently scrollable, the entire page is now scrollable when the content's height exceeds the height of the viewport. The sidebar is sticky and will remain at the top of the viewport while scrolling. The infobar is similar, but if its content's height exceeds the height of the viewport, it will scroll alongside the viewport until it reaches its bottom, at which point it will stick. This is different to the sidebar, which will only fill the available viewport height and will become scrollable itself if its height exceeds the height of the viewport.