Write utility function for validating uploaded files
This function is tied to the File input component as it validates files read as array buffers and converts the uploaded data to the type sent from the component's onUpdateHandler callback. I have also updated the component itself to use this new utility function, and export the 'FileInfoTypes' type for the utility function to use.