Add 'disabled' prop to Upload button element
This prop *did* exist on the component, but it was set through the Basic button props supplied to the component. Now, the prop is at the 'top-level' of the component, which is more appropriate given its relationship to the component; it is supposed to disable the input element component and not just the Basic button component it utilises.