Change current post routes
Instead of the routes /post/create and /post/delete?postId for the post operations, the paths for these routes are now /post and /post/:postId, respectively, using the correct HTTP verb to designate which route performs which operation - this is more CRUD-like. The DELETE route now also requires a path parameter to specify the specific post to delete rather than a query parameter.