Wrap Main profile component in context provider
I have done this to fetch the _id of the user whose profile being viewed only once, so it can be provided to other profile feature components, instead of having to fetch it from each one. I am now also importing the UserPosts component and am using it in-place of the placeholder content for the /profile/:accountTag route.