Fix issue in UserPosts and UserLikes components
The parameters supplied to the useAsync hooks in each component were not being reset after a change in the _id provided by the ProfileContext, meaning the api functions weren't being called with a valid MongoDB ObjectId, so an error was always returned until the user switched tabs in the Main profile component and switched back to trigger a rerender after the user's _id had been fetched. This has now been fixed; the components react to a change in the user's _id value when it has been fetched successfully.