Update post List component
I've added a few features into this component to make it mimic the UserPosts profile component (the error message element, 'empty' message element, it now invokes an API function and utilises a useAsync hook, etc.). As it stands, the new 'getRecommendedPosts' function is just a copy of the 'getUserPosts' function - I am still yet to write a route on the backend that fetches posts from followed users, so this function is just a placeholder for now.