Change how following a user is handled
Before, following a user would cause the component to attempt to fetch the current user summary information from the server. However, this would cause the skeleton display to appear, which wasn't very visually appealing. Now, the 'Follow'/'Unfollow' button is updated accordingly with respect to the status of the response from attempting to follow/unfollow the user, and the user's 'followers' count is also updated. In addition to this, the 'Follow'/'Unfollow' button itself is now disabled when awaiting a response from the server when attempting to follow/unfollow the user, preventing multiple calls to the API.