Perform complete overhaul of chat Option component
The component now utilises the 'getChatOverview' function to fetch the necessary information for the chat (the ObjectId reference for which is specified with the _id prop). Also, clicking a chat will now redirect the user to the correct URL path for that chat. I have also implemented the Skeleton component for some visual effects to convey the information that the component is awaiting a response from the server containing the information to be displayed for the chat.