Overhaul chat Message component
The component now utilises the 'getChatMessage' function to fetch the information for the messageId provided as a prop, and displays a 'skeleton page' effect for the message while waiting. When the data has been received, the correct information is displayed for that message and the message it is replying to, if applicable.