Overhaul Active chat component
The component is now a context provider that contains some important information about the active chat including the name, and participants' information. It utilises the 'getChatOverview' function to fetch this information from the server. I have also moved the 'message list' part of the component to its own component (MessageList).