Utilise updateChatName function in chat Header component
A request can now be made for the chat's name to be changed from the component. I have also changed the options when the name is being edited; there are now two buttons: a 'done' button and a 'close' button. An error message will now display if there are any errors as a result of the request. The style rules for the component have also been updated.