Fix issue in chat Header component
After changing the chat name once, every time the chat name had a single character changed, the 'editingName' state bool was being set to false, because the useEffect hook checked to see if a response had been obtained from the server and, if so, was setting 'editingName' to false. By storing the response in state and setting it to 'null' after the initial conditional check, it no longer sets 'editingName' to false until another response from the server is received, updating that stored state, which fixes the issue.