Write custom hook for scrollable elements
The hook can be passed two callback functions that are called when either the scrollbar of the scrollable element reaches the top or the bottom of its possible range. It also stores and returns state about whether the element is currently scrolled to the top or bottom of its scrollable range.