Write functionality for requesting additional posts when scrolling
I have also adjusted the layout for the error message and create new post button elements such that there is available space for them at the top and bottom of the posts list, respectively. For example, the error message is displayed at the top of the list, but since it is sticky, it covers the top of the first post in the list, so I have created a buffer equal to the size of the element at the top of the list for it to be positioned when at the top of the scrollable area. This way, it will not cover the first post in the list. I've done the same for the create new post button and the last post in the list.