Fix similar problem on various /user GET route middlewares
Firstly, I moved the  part of the aggregation pipeline that matches the specified 'after' post/chat/liked post (if applicable) to earlier in the pipeline. When there were no older posts/chats/liked posts than the specified post/chat/liked post ('after' URL param), the aggregation was returning nothing - I wanted an empty array of those documents to be returned by the query. Replacing the  stage of the pipeline that filtered those documents with an actual  solved this issue.