Write functionality for requesting additional post likes when scrolling
Similarly to other list components, I have updated the layout for the error message. I have also made the 'Return to post...' button sticky so that, no matter how far down the list of likes the user has scrolled, they can easily click it to navigate to the parent post.