Overhaul Post component
The component now longer displays replies as children of itself - this functionality will now be handled by a different component. The reason I have decided to do this is because the Post component was becoming increasingly convoluted; I thought it would be better to move some of the responsiblities to another component. I have also updated the component's layout so the error message, when displayed, appears alongside the Post (before, the error would not display if the post had loaded, which meant any errors caused by liking the post would not be visible). Also, the 'Like' button is now disabled whenever the component is awaiting a response from the server as a result of attempting to like/dislike the post.