Fix problem on various /post GET route middlewares
In the event that no likes or replies were matched after the specified 'after' URL parameter, the pipeline was returning nothing. I wanted it to return an empty array to indicate the post itself had been found, but the specified parameters caused the likes/replies arrays to be empty. This has now been fixed.