Write functionality for handling successful replies to post
By subscribing to the topic published when a post is successfully created, the Replies component now checks to see if the 'replyingTo' field for the new post matches its own id. If so, the new post is added to the start of its 'postReplies' state array. If it is instead a reply to one of its replies, the user is redirected to the page displaying the post that was replied to, and the new post will be displayed as a reply of that post. I have also added a way to forcibly rerender and reset the component when using the useNavigate hook inside the Post component, as it wasn't triggering a rerender when the URL path changed. The Replies component now reacts to a change in the URL path and updates accordingly.