Store server response in state
Without this, the 'fieldInitialValue' state was being set to the current value of the field on every update of the field's value after the first successful request. Storing the response's state and setting the state to 'null' after reading its contents allows the component to set the initial value of the field only once after each successful request.