Fix error message issue in various list-based components
When the error message was being cleared, the space made available for it at the top of some list-based components was remaining, as the height was based on the height of the ref element. However, the ref was not being set to the empty error message element, so its height wasn't being reset to 0 when the error message was cleared. This issue is now fixed.