Update infobar Wrapper component to be customisable
The chosen panels to display in the infobar in addition to any children can now be set using the PubSub package. The reason I have done this is to allow for the sidebar to change from anywhere in the component heirarchy while mounting the Wrapper component itself at a much higher level. Before, the Wrapper component was being used in various nested components that, when mounted/unmounted, would trigger a rerender in the sidebar - I wanted the sidebar to change without rerendering panels that stayed consistent between different pages, and this method achieves this.