Publish 'infobar-ready' topic on mount
Due to the rendering order (the infobar being one of the last components rendered), other components that were publishing to topics on mount for the infobar Wrapper component were doing so before the Wrapper had subscribed to those topics. By publishing the 'infobar-ready' topic, other components will be able to subscribe to this topic to know when the Wrapper component is ready to receive data for topics it is subscribed to. I have also changed the name of the topics that used 'sidebar' in them, as this is an infobar component, not a sidebar component.