Fix appearance when awaiting API response
The 'no userId provided...' message is now no longer displayed while the UserContext component is awaiting a response. Also, the 'No recent posts to display' message is not displayed unless the posts have been fetched and the array is definitely empty. The WaitingWheel accessibility component is now only displayed when waiting for a response and when the posts array is either falsy or empty.