update /user/:userId/chat-activity GET route's middleware
I wanted the pipeline to maintain the structure of its returned information in the case where there were no chats with any activity that satisfied the conditions for them to be considered 'recent'. Before, it was simply returning either an array of chats or an empty array, but now it returns an array with one object that contains a 'populatedChats' array field with the valid chats. This is more useful than the chats being returned in the base array because now I can more easily distinguish between when the user's credentials were satisfactory to the point the aggregation completed but there were no valid chats (returns the object) and when the pipeline failed for another reason like a failed  stage (returns an empty array).