Update useScrollableElement hook
My goal was to provide some additional context along with this hook in the form of the state variable: 'isScrollable', which would state whether the element is scrollable. To achieve this, I have changed how the hook works by using a ref callback within the hook instead of providing the ref to  the hook from the component utilising the hook. This callback is then provided as the value to the 'ref' property for the desired scrollable element. I have also started returning context from the hook in object format instead of as an array, to make selective destructuring possible.