Load more posts when there is still available space in container
By default, a limit of 10 posts are loaded at a time and more are loaded when the user scrolls to the bottom of the container. But if those 10 posts don't fill the container for some reason (the user may have a very tall monitor or may be zoomed out), and there are still more posts in the database that can be loaded, subsequent requests will be made for more posts until either the space is filled or no more posts can be loaded.